Current worst problem: SLOW
incubot is taking about _ten seconds_ to come up with a witticism.
Most of the time seems to be spent computing the number of occurences
of each word in the utterance that triggered it.  I wonder if I could
use some sort of index on the log_word_map table.