Idea for a worthwhile optimization
Finding the popularity of a word takes time proportional to that
popularity.  For example, to find the popularity of the word "it", we
must scan through the log_word_map index, counting the number of
occurrences.  For a common word like "it", that's slow!

Since we're ultimately only interested in _unpopular_ words, we can
simply filter out a bunch of the more popular words from our input
set, and rank only the remaining words.  This little example shows
that without such filtering, it takes about 8 seconds to return a
witticism; with filtering, it's .4 seconds.