Fix highly theoretical issue
PLEASE NOTE THAT I DID NOT TEST THIS CODE AT ALL, EVEN WITH CHECK SYNTAX

With the current code, if the client thread receives a value in its mailbox while waiting on the incubot-server, it could return a wrong result. Passing in an async-channel to the server should fix this.