Repro the bug, maybe
The behavior "in production" was different: rather than simply
displaying "out of time", it died with an ABORT signal.  But I suspect
it's due to running out of memory, and I also suspect the cure is to
simply reduce the custodian's memory limit, so that we hit _that_ limit
before the kernel is provoked to kill us.